Undo changes not related to rebind sample extension.
