--- Packages/com.unity.inputsystem/InputSystem/Plugins/OnScreen/OnScreenHelpers.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: b04044ee8dfc48409d1d55a48af5cb77
-timeCreated: 1758782539
